feat(10): add protected routes
Add the ability to create protected routes. They will redirect to login and you will be returned to the page you were on after logging in.

There are some issues with the login flow running after reloading the page but it might be due to running it locally. Something to investigate later.